added chunk and document and converted them into pydantic models
