+++ src/vectordb/splitter/interface.py
@@ -1,5 +1,5 @@
 
-class Chunker(VerbaComponent):
+class Chunker(Component):
     """
     Interface for Verba Chunking.
     """
